+++ src/api/event.api.js
@@ -1,4 +1,40 @@
+import { get } from 'lodash';
+import { waterfall, parallel } from 'async';
+import { idOf, mergeObjects } from '@lykmapipo/common';
+import { Predefine } from '@lykmapipo/predefine';
 import Event from '../event.model';
+import { sendEventNotification } from './notification.api';
+
+export const preLoadRelated = (optns, done) => {
+  // ensure options
+  const options = mergeObjects(optns);
+
+  const ensureType = next => {
+    const typeId = idOf(options.type) || options.type;
+    if (typeId) {
+      // TODO: or find default
+      return Predefine.getById({ _id: typeId }, next);
+    }
+    return next(null, typeId);
+  };
+
+  const ensureGroup = next => {
+    const typeId = idOf(options.type) || options.type;
+    if (typeId) {
+      // TODO: or find default
+      // TODO: try use group if exists
+      return Predefine.getById({ _id: typeId }, (error, type) => {
+        const group = get(type, 'relations.group');
+        return next(error, group);
+      });
+    }
+    return next(null, null);
+  };
+
+  // execute tasks
+  const tasks = { type: ensureType, group: ensureGroup };
+  return parallel(tasks, done);
+};
 
 export const getEventJsonSchema = (optns, done) => {
   const jsonSchema = Event.jsonSchema();
@@ -20,7 +56,27 @@ export const getEventById = (optns, done) => {
 };
 
 export const postEventWithChanges = (optns, done) => {
-  return Event.post(optns, done);
+  const options = mergeObjects(optns);
+
+  return waterfall(
+    [
+      next => preLoadRelated(options, next),
+      (related, next) => {
+        const event = mergeObjects(options, related);
+        return Event.post(event, next);
+      },
+      (event, next) => {
+        // TODO: check AUTO_EVENT_NOTIFICATION_ENABLED=true
+        return sendEventNotification(event, (/* error, sent */) => {
+          // TODO: notify(or log) swallowed error
+          return next(null, event);
+        });
+      },
+      // TODO: ensure level, severity, certainty, status, urgency
+      // TODO: save initial changelog
+    ],
+    done
+  );
 };
 
 export const putEventWithChanges = (optns, done) => {