+++ test/integration/notification.api.spec.js
@@ -1,0 +1,54 @@
+import { clear, create, expect } from '@lykmapipo/mongoose-test-helpers';
+import { Predefine } from '@lykmapipo/predefine';
+import { Event } from '../../src';
+
+import {
+  CHANNEL_EMAIL,
+  ENABLE_SYNC_TRANSPORT,
+  NOTIFICATION_CHANNELS,
+  TEMPLATES_EVENT_NOTIFICATION_TITLE,
+  TEMPLATES_EVENT_NOTIFICATION_MESSAGE,
+  Campaign,
+  sendCampaign,
+  sendEventNotification,
+} from '../../src/api/notification.api';
+
+describe('Notification', () => {
+  const level = Predefine.fake();
+  const type = Predefine.fake();
+  const areas = Predefine.fake(2);
+
+  const event = Event.fake();
+  event.set({ level, type, areas });
+
+  before(done => clear(done));
+  before(done => create([level, type, ...areas], done));
+  before(done => create(event, done));
+
+  it('should parse & expose common config', () => {
+    expect(CHANNEL_EMAIL).to.exist;
+    expect(ENABLE_SYNC_TRANSPORT).to.exist;
+    expect(NOTIFICATION_CHANNELS).to.exist;
+    expect(TEMPLATES_EVENT_NOTIFICATION_TITLE).to.exist;
+    expect(TEMPLATES_EVENT_NOTIFICATION_MESSAGE).to.exist;
+  });
+
+  it('should send a campaign', done => {
+    const message = Campaign.fake().toObject();
+    sendCampaign(message, (error, sent) => {
+      expect(error).to.not.exist;
+      expect(sent).to.exist;
+      done(error, sent);
+    });
+  });
+
+  it('should send event alert', done => {
+    sendEventNotification(event, (error, sent) => {
+      expect(error).to.not.exist;
+      expect(sent).to.exist;
+      done(error, sent);
+    });
+  });
+
+  after(done => clear(done));
+});